backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
torch==2.2.0
torchvision==0.17.0
transformers==4.35.0
openai==0.28.1
langchain==0.0.340
chromadb==0.4.18
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
pydantic-settings>=2.0.0